Refactor: move REPL and main.rs into crates/database-cli
Move src/main.rs and src/repl/ into crates/database-cli/src/, removing them from
the root library package. Add REPL/TUI dependencies (rustyline, ratatui, crossterm,
ansi-to-tui, colored, rand, ciborium) to the CLI crate so it builds independently.
The root package no longer has a binary target.
