+++ crates/database-cli/src/repl/shared.rs
